add client tool kill rank subcommand
